Started the implementation of a manual version of a langchain agent
I had a lot of small issues with langchain and decided it might be worth
trying to make a fully custom agent. This would also allow the few-shot
prompt texts to be more specific to the task at hand and hopefully guide
the agent further.

Further, in order to get around the openai ethics guidelines, I have
turned the example text into a game. The whole time the agent is
attempting to hack devices on a network, it will think that its playing
a text-based hacking game.

This idea is not yet finished and not tested.