Re inventing the wheel
The langchain framework is either not very configurable, or I am
struggling to find the right configurations. A lot of the times the
documentation is insufficient and I have to read the code for the
framework to actually know what to do.

The combination of the code being relatively simple, and my endless
frustration with langchain lead me down the road of implementing the
agents part of langchain on my own to serve my needs. This is what has
been added in this commit, under research/reinvent-the-wheel