Progress made on setting the agent up to make it be capable of executing more complex tasks.
I have added a small entity database like knowledge base system to help
keep track of useful information during multiple runs of the agent. The
knowledge base (called entityDB) keeps track of all hosts and targets
that we have discovered. It keeps in plain text some amount of detail
about each target.

The database will be used both as a tool which the agent can call upon
to get information about specific things and as a 'sort of' vector
database that automatically gives some context to a task.

The database is not yet stable (and likely never be considering its
controlled by language models), but its at a stable enough state to move
on with other aspects of the project.

Additionally to the database, I have added a small test program called
hacker.py that I'm using to continuously test the capabilities of the
agent at hacking as I'm adding more features.