Re-implementing langchains shelltool with sandboxing and bug fixes
On my continuous quest to re-implement all of langchain (joke), I have
now done most of the shelltool module.

My issue of with the current langchain shelltool is that it just runs
command on the main system without any sandboxing. If the agent decides
to install programs it cannot do so as it does not have root, and giving
the agent root access seems to be an unnecessary risk. Further, with or
without root, I do not trust the agent to not create problems on my
computer.

My re-implemented version uses podman (or docker) containers to execute
code. The shelltool program is complete, however it is not hooked up
with the agent yet as I want to make sure it cannot be exploited to run
commands outside the sandbox.

Finally I have also fixed some small parsing bugs on the agent itself (as usual
lately)