Added a simple command filter to the shelltool.
Since the commands are being run in a container, and the LLM will not
know this. I have added a small filter that stops it from using sudo or
apt without the -y option. These small filters might fix some small
common issues, but larger ones coming from the fact that it cannot
interact with the terminal might not be so easy to fix.

Additionally, the container does not have vim, or any text editor which
might force it to use echo instead.