More work done on the tasker
The goal of the tasker has been changed slightly to be a lot more
hard-coded.

Originally the tasker was supposed to generate all the tasks in real
time, however, since the actual system does not need to be that flexible
in what it has to do we can hardcode some of the tasks to make it a lot
more stable.

The new tasker should have a combination of hardcoded tasks, templated
tasks and maybe some generated ones in the future.