Testing and fixing small issues
To test the all the different libraries that have been created working
together, I'm currently making the hacker.ipynb notebook. Essentially in
this notebook I start writing the code for the full hacking and scanning
agent and I improve / fix bugs in the libraries as they come up. This
way I always know what its most important to fix.

In this commit there have been some minor changes and bug fixes,
including, but not limited to agents now having a context field where
the database can put its context for the task.

Further I came across a bug with the networking when using docker
containers. At this point I do not know if its an issue with my docker
setup or my computer as a whole and I have left it unfixed for now. More
details on it can be found in a large comment on shelltool.py above
where the docker container is started.