Workaround for docker bug
Unfortunately the bug where docker can either use nmap or have --network
host set to have the correct IP address could not be fixed. Not sure if
this is a limitation of docker/podman or a skill issue.

However, I have come up with a slightly cursed workaround. If the Ip a
or ifconfig command is run, the command will run on host instead of the
sandbox. Further, there is now a way to set other commands to do this if
needed int he future.

Additionally, this commit also adds a patch to the nmap command. If the
agent chooses to run `nmap -sn`, the command will be replaced by `nmap
-PO -sn` as raw packets work better within the docker container.

This commit fixes all known issues with the shelltool, thus development
of the tasker can continue.