Work towards migrating to the new openai client
This commit migrates the agent file to the new openai library as well as
the main functions in other files. Other files have not been migrated
yet, including the entityDB file which has many calls to openai. As
these were done with an instance variable holding the openai client, the
automatic migrate function did not manage to address them.

Further the migrate function did not change any part where the return of
the model was turned into text (many changes here).