Made progress on task manager and new database models
Currently everything should work, minus one bug I do not have time to
fix now. The bug is simply that everytime the Database gets
re-initialized, the target list gets set to empty. This makes sense as
it is defined in the __init__ of the Database().

I need to create some sort of system to stop this happening, probably
separating the target list from the db in some way. I don't have time to
do this now, but should be a simple task whenever I get back to this
work