Re-built the tasker to incorporate both the target list and tasks
Tasks are now sorted by target so that we can run tasks sequentially
without having to wait for the timout of the target.

The new task manager file is class has been implemented and roughly
tested, however it has not yet been connected up to the rest of the
project. This will be done after more testing