Working on the c2 server
Currently I have a basic c2 server and a client called c2communication.
The client and the server only have the update target list part set but
I cannot test it as the starter pinecone instance went down. I will
have to come back to test it later.

The main thing that needs testing is whether the singleton task_manager
works well enough that I can get live targets in the communication file
without having to pass in the task manager. If this doesn't work then i
have to pass it in.