Updates to c2 server and communication
In the first iteration of the c2 server, it was just storing lists of
targets and the blacklist. Each list only contained the name of the
target.

This was changed today to hold the entirety of the target_list with all
its attributes. Further a stopped variable was added to the target_list,
that when set to true, will be ignored. This paves the way for a user
being able to exclude a target from scans.

Some other work has been done the user-side of the server, however no
feature is complete yet.

Lastly, there were a few failsafes added to the agent.
1. pinecone is unstable lately, so each pinecone request includes a
   try:except: block with a 2 second timeout to try again.
2. Sometimes when setting targets as stopped, the get_next task finder
   thinks that there are no more targets available. A failsafe has been
   added here to add a general task to re-scan for new devices. If some
   are found they get added to target list, if none are found then it
   should continue on with the devices there are. If there are no
   devices left it will keep re-scanning until it finds one or the user
   kills the agent
3. auto_mark_completed was changed to take in a target and mark their
   next task as completed instead of figuring out the target. This is
   because if a target is stopped it could mess up the whole task
   manager

BUGS: There is an interesting behaviour however, that when there is one
target set as stopped and there is only one other target, than it seems
to hit the general network scan on every other command. If there is only
one target, then this is not a bad behaviour, however it should not do
this if there are more. I have not had more targets to test this, but
hopefully it doesnt