Working on website for c2 server
Due to time limitations I have chosen to not do an android application
for now, but rather a website where you can control the devices. This
commit makes good progress on the website, with all instances displayed,
and within instances each target with all tasks.

Features missing are the following:
- making website interactive
- making report and pause buttons work
- having debug data on the c2 server
- some sort of extra information about the target on the page