Working on c2 server website
This commit finishes most of the user facing code other than
authentication.

Currently features that work are:
- displaying all instances
- displaying all devices in an instance
- dropdown menu with tasklist showing completed and not completed
- working generate report button, although it is not the best quality
  report
- working pause/resume button on devices

Next things that need to be implemented:
- breaking appart the report model into many smaller parts
Currently a single model does the entire report, however this is not
that good. All data the model collects comes down to this report, thus
the report must be of top quality. I think it should be split into each
subsection and a seperate section for each port. Or maybe some more
dynamic way of splitting into subsections.
- Admin authentication.
This should be simple as we can just define a username and password in
the .env file hashed