Finished basic authentication for the c2 server
I have created a very simple authentication system for the admin to
check on the c2 server. I need this site to be very secure, but wanted
to keep it simple, so there is only one user and there is no way to
register more. The username and password are hashed and stored in
environment variables, and a single cookie is stored as a varaible in
the code (only one device can be logged in at once).